+++ Sources/SyntaxAnalysis/ImportSyntaxVisitor.swift
@@ -14,12 +14,18 @@ public final class ImportSyntaxVisitor: PeripherySyntaxVisitor {
     public func visit(_ node: ImportDeclSyntax) {
         let parts = node.path.map(\.name.text)
         let module = parts.first ?? ""
-        let attributes = node.attributes.compactMap { $0.as(AttributeSyntax.self)?.attributeName.trimmedDescription }
+        let attributes = node.attributes.compactMap {
+            if case let .attribute(attr) = $0 {
+              attr.attributeName.trimmedDescription
+            } else {
+                nil
+            }
+        }
         let location = sourceLocationBuilder.location(at: node.positionAfterSkippingLeadingTrivia)
         let statement = ImportStatement(
             module: module,
             isTestable: attributes.contains("testable"),
-            isExported: attributes.contains("_exported"),
+            isExported: attributes.contains("_exported") || node.modifiers.contains { $0.name.text == "public" },
             location: location
         )
         importStatements.append(statement)