+++ Sources/PeripheryKit/SourceGraph/SourceGraphMutatorRunner.swift
@@ -44,6 +44,7 @@ public final class SourceGraphMutatorRunner {
         ResultBuilderRetainer.self,
         CapitalSelfFunctionCallRetainer.self,
         CodablePropertyRetainer.self,
+        ExternalOverrideRetainer.self,
 
         AncestralReferenceEliminator.self,
         AssignOnlyPropertyReferenceEliminator.self,